Rename zwave_js trigger helper to avoid confusion
