+++ homeassistant/components/zwave_js/triggers/value_updated.py
@@ -30,7 +30,7 @@ from ..const import (
     DOMAIN,
 )
 from ..helpers import async_get_nodes_from_targets, get_device_id
-from .helpers import async_bypass_dynamic_config_validation
+from .trigger_helpers import async_bypass_dynamic_config_validation
 
 # Platform type should be <DOMAIN>.<SUBMODULE_NAME>
 PLATFORM_TYPE = f"{DOMAIN}.{__name__.rsplit('.', maxsplit=1)[-1]}"
